Python/requirements.txt
beautifulsoup4==4.12.3
bibtexparser==2.0.0b7
Requests==2.31.0
langchain_community==0.2.6
langchain_openai==0.1.9
langgraph==0.0.69
psycopg2-binary==2.9.9
python-dotenv==1.0.1
types-beautifulsoup4==4.12.0.20240511
html2text==2024.2.26
pandas==2.2.2
